requirements.txt
pybind11
pyyaml
aiosmtplib
uvloop